2007-09-11  Dan Williams  <dcbw@redhat.com>
	* libnm-glib/nm-access-point.c
	  libnm-glib/nm-access-point.c
		- (nm_access_point_get_frequency): now returns guint32 to match
			property change on 2007-09-10



git-svn-id: http://svn-archive.gnome.org/svn/NetworkManager/trunk@2788 4912f4e0-d625-0410-9fb7-b9a5a253dbdc

+++ libnm-glib/nm-access-point.h
@@ -33,7 +33,7 @@ guint32      nm_access_point_get_flags        (NMAccessPoint *ap);
 guint32      nm_access_point_get_wpa_flags    (NMAccessPoint *ap);
 guint32      nm_access_point_get_rsn_flags    (NMAccessPoint *ap);
 GByteArray * nm_access_point_get_ssid         (NMAccessPoint *ap);
-gdouble      nm_access_point_get_frequency    (NMAccessPoint *ap);
+guint32      nm_access_point_get_frequency    (NMAccessPoint *ap);
 char *       nm_access_point_get_hw_address   (NMAccessPoint *ap);
 int          nm_access_point_get_mode         (NMAccessPoint *ap);
 guint32      nm_access_point_get_rate         (NMAccessPoint *ap);
